Safari-609.1.11.1/LayoutTests/fast/forms/option-constructor-selected-expected.txt

This tests that the "selected" argument is honored when constructing an option element.

Test 0 Passed: the selected index is 0.
Test 1 Passed: the selected index is 0.
Test 2 Passed: the selected index is 0.
Test 3 Passed: the selected index is 3.
Test 4 Passed: the selected index is 4.
Test 5 Passed: the selected index is 4.
